# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: ploykit API
  version: 0.3.0
  description: |
    ploykit 框架 + example 产品应用的 HTTP 契约。

    认证两通道（webx.Authenticate）：
    - 浏览器会话 Cookie `tk_auth`（写请求须带 X-CSRF-Token，/config 拉取 masked token）；
    - PAT `Authorization: Bearer tk_*`（豁免 CSRF）。

    扁平路由（/api/billing/*、/api/usage、/api/audit*、/api/tasks*）的工作区上下文
    由 `X-Workspace-Id` 请求头注入（@ploykit/client 自动携带）。

    错误信封统一为 ErrorBody（webx.ErrorBody）：`{"error":"E_*","message":string,"details"?}`。
security:
  - sessionCookie: []
  - patBearer: []
...
paths:
...
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [account]
      summary: 凭邮件令牌重置密码（成功后旧会话全吊销）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/send-verification:
    post:
      operationId: sendVerification
      tags: [account]
      summary: 发送邮箱验证邮件（仅浏览器会话）
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 已发送
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/DefaultError'
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [account]
      summary: 凭令牌验证邮箱
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: 验证成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- identity：会话与资料 ----------
...
  /api/audit:
    get:
      operationId: listAuditEvents
      tags: [audit]
      summary: 当前工作区审计事件（audit:read；X-Workspace-Id；created_at 倒序）
      parameters:
        - $ref: '#/components/parameters/AuditActorID'
        - $ref: '#/components/parameters/AuditAction'
        - $ref: '#/components/parameters/AuditResourceType'
        - $ref: '#/components/parameters/AuditFrom'
        - $ref: '#/components/parameters/AuditTo'
        - $ref: '#/components/parameters/AuditLimit'
        - $ref: '#/components/parameters/AuditOffset'
      responses:
        '200':
          description: 审计事件 + 过滤命中总数
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'
  /api/audit/export.csv:
    get:
      operationId: exportAuditCsv
      tags: [audit]
      summary: 当前工作区审计 CSV 导出（同 list 过滤参数；单次查询内存缓冲，上限 5000 行，limit 超上界按 5000 截断且无截断标记头——与 admin 导出的 50000 行流式/截断头是两套实现）
      parameters:
        - $ref: '#/components/parameters/AuditActorID'
        - $ref: '#/components/parameters/AuditAction'
        - $ref: '#/components/parameters/AuditResourceType'
        - $ref: '#/components/parameters/AuditFrom'
        - $ref: '#/components/parameters/AuditTo'
        - $ref: '#/components/parameters/AuditLimit'
        - $ref: '#/components/parameters/AuditOffset'
      responses:
        '200':
          description: CSV 附件（列头 created_at,actor_type,actor_id,actor_email,action,resource_type,resource_id,request_id,metadata）
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---------- admin（platform admin） ----------
...
  /api/admin/users/{id}:
    get:
      operationId: adminGetUserDetail
      tags: [admin]
      summary: 用户详情聚合（资料+在途会话+PAT+所属工作区；platform admin）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 详情聚合 }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: 软删账号（?confirm=<email> 二次确认；不可删自己/另一平台管理员）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: confirm, in: query, required: false, schema: { type: string }, description: 目标邮箱二次确认 }
      responses:
        '200':
          description: 已软删（status=deleted+邮箱改写+PAT 全吊销）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        '400': { $ref: '#/components/responses/DefaultError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/kick:
    post:
      operationId: adminKickUser
      tags: [admin]
      summary: 踢下线（吊销目标全部在途会话）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已吊销
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        '503': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/sessions/{sid}/revoke:
    post:
      operationId: adminRevokeUserSession
      tags: [admin]
      summary: 吊销单个会话（不存在/已吊销/非本人会话 → 404 幂等）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: sid, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已吊销
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        '404': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/password-reset:
    post:
      operationId: adminResetUserPassword
      tags: [admin]
      summary: 触发密码重置邮件（复用自助流：防枚举/1 分钟冷却）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已发送
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/resend-verification:
    post:
      operationId: adminResendVerification
      tags: [admin]
      summary: 补发邮箱验证邮件（已验证 409 / 冷却 429）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已发送
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        '409': { $ref: '#/components/responses/DefaultError' }
        '429': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/mark-verified:
    post:
      operationId: adminMarkVerified
      tags: [admin]
      summary: 直接标记邮箱已验证（幂等）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已标记
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/pats:
    get:
      operationId: adminListUserPATs
      tags: [admin]
      summary: 目标用户 PAT 元数据列表（只读，不含凭据材料）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: PAT 列表 }
        default: { $ref: '#/components/responses/DefaultError' }
  /api/admin/users/{id}/pats/{patID}/revoke:
    post:
      operationId: adminRevokeUserPAT
      tags: [admin]
      summary: 吊销目标用户 PAT（404 幂等）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: patID, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已吊销
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        '404': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
...
  /api/admin/users/{id}/export:
    get:
      operationId: adminExportUserdata
      tags: [admin]
      summary: GDPR 用户数据导出（JSON 附件；password_hash/token_hash 类字段源头不选）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 聚合数据 JSON 附件（filename 为 userdata-{id}.json；body 键含 user/memberships/sessions/personal_access_tokens/audit_events/exported_at）
          content:
            application/json: {}
        '404': { $ref: '#/components/responses/DefaultError' }
        default: { $ref: '#/components/responses/DefaultError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tk_auth
      description: 浏览器会话 Cookie（webx.Authenticate 签发/续期；写请求须另带 X-CSRF-Token）
    patBearer:
      type: http
      scheme: bearer
      description: '个人访问令牌（Authorization: Bearer tk_*，CSRF 豁免）'
  parameters:
...
    AuditActorID:
      name: actor_id
      in: query
      required: false
      schema: { type: string }
      description: 过滤：操作者 ID
    AuditAction:
      name: action
      in: query
      required: false
      schema: { type: string }
      description: 过滤：动作（如 workspace.rename）
    AuditResourceType:
      name: resource_type
      in: query
      required: false
      schema: { type: string }
      description: 过滤：资源类型
    AuditFrom:
      name: from
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: 过滤：created_at >= from（RFC3339）
    AuditTo:
      name: to
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: 过滤：created_at < to（RFC3339）
    AuditLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
      description: 每页条数（1..500，缺省 100）
    AuditOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 偏移
  responses:
    DefaultError:
      description: 其他错误（统一 ErrorBody 信封）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadRequest:
      description: 参数非法（E_VALIDATION / E_BAD_JSON / E_PAYLOAD_TOO_LARGE）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthenticated:
      description: 未认证（E_UNAUTHENTICATED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: 无权限（E_FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    RateLimited:
      description: 限流（E_RATE_LIMITED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 错误码（E_UNAUTHENTICATED / E_FORBIDDEN / E_NOT_FOUND / E_VALIDATION / E_CONFLICT / E_RATE_LIMITED / E_QUOTA_EXCEEDED / E_INTERNAL / E_TIMEOUT / E_PAYLOAD_TOO_LARGE / E_UNAVAILABLE）
        message:
          type: string
        details:
          description: 可选补充（通常为字符串详情）
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    ResetPasswordRequest:
      type: object
      required: [email, token, new_password]
      properties:
        email:
          type: string
        token:
          type: string
        new_password:
          type: string
    VerifyEmailRequest:
      type: object
      required: [email, token]
      properties:
        email:
          type: string
        token:
          type: string
...
    AuditEvent:
      type: object
      required: [id, workspace_id, actor_type, actor_id, actor_snapshot, action, resource_type, resource_id, metadata, created_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
          description: 工作区 ID（平台级事件为空串）
        actor_type:
          type: string
          enum: [user, pat, system]
        actor_id:
          type: [string, 'null']
          description: 操作者 ID（system 为 null）
        actor_snapshot:
          type: object
          additionalProperties: true
          description: 操作时身份快照（email / role / workspace_id）
        action:
          type: string
          description: 动作（如 workspace.rename / billing.plan_changed）
        resource_type:
          type: string
        resource_id:
          type: [string, 'null']
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        total:
          type: integer
          description: 过滤命中总数（非当页条数）

    # ---- admin ----
...